requirements.txt
python-dotenv
langchain
langchain_openai
langchain_pinecone
langchain_community
openai
chromadb
tiktoken
streamlit
pinecone-client
pypdf
pysqlite3-binary